Some bugs fixed and added README
