+++ Analysis/include/MvaMethods.h
@@ -5,8 +5,11 @@ This file is part of https://github.com/hh-italian-group/hh-bbtautau. */
 #pragma once
 
 #include "hh-bbtautau/Analysis/include/MvaConfiguration.h"
+#include "hh-bbtautau/Analysis/include/MvaVariablesStudy.h"
 #include "AnalysisTools/Core/include/RootExt.h"
 #include <TCanvas.h>
+#include <future>
+#include "AnalysisTools/Run/include/MultiThread.h"
 
 namespace  analysis {
 namespace mva_study{
