+++ Studies/config/mva_config.cfg
@@ -1,77 +1,36 @@
-[LowMassRange]
+[Common]
 channels: eTau muTau
+param_list: BoostType Grad
+param: V False
+param: UseBaggedBoost True
+param: nCuts 20
+param: NTrees 400
+param: shrinkage 0.1
+param: BaggedSampleFraction 0.6
+# NodePurityLimit
+# Pruning ??
+# param_range: nCuts 20:100:80
+######
+#param_range: NTrees 300:1200:300
+#param_range: shrinkage 0.1:1.0:0.3
+#param_range: BaggedSampleFraction 0.5:1.0:0.25
+#param_range: MaxDepth 2:5:1
+#param_range: MinNodeSize 0.01:0.09:0.04
+######
+significant_params: BoostType
+use_mass_var: true
+
+[Low : Common]
 mass_range: 250 320
-param_list: BoostType Grad
-param: V False
-param: UseBaggedBoost True
-param: nCuts 20
-param: NTrees 400
-param: shrinkage 0.1
-# NodePurityLimit
-# Pruning ??
-# param_range: nCuts 20:100:80
-######
-param_range: NTrees 100:900:400
-param_range: shrinkage 0.1:1.0:0.3
-param_range: BaggedSampleFraction 0.5:1.0:0.25
-param_range: MaxDepth 2:4:1
-param_range: MinNodeSize 0.01:0.09:0.04
-######
-significant_params: BoostType
 variables: HT_otherjets MT2 MT_htautau MT_l1 MT_tot abs_dphi_l1MET abs_dphi_htautauMET dR_b1b2_boosted dR_l1l2Pt_htautau dR_l1l2_boosted dphi_b1b2 dphi_hbbhtautau dphi_l2MET mass_l1l2MET mass_top1 mass_top2 p_zeta p_zetavisible pt_hbb pt_htautau pt_l1 pt_l1l2MET pt_l2
-use_mass_var: true
 
-
-
-[MediumMassRange]
-channels: eTau muTau
+[Medium : Common]
 mass_range: 340 400
-param_list: BoostType Grad
-param: V False
-param: UseBaggedBoost True
-param: nCuts 20
-param: NTrees 400
-param: shrinkage 0.1
-# NodePurityLimit
-# Pruning ??
-# param_range: nCuts 20:100:80
-######
-param_range: NTrees 100:900:400
-param_range: shrinkage 0.1:1.0:0.3
-param_range: BaggedSampleFraction 0.5:1.0:0.25
-param_range: MaxDepth 2:4:1
-param_range: MinNodeSize 0.01:0.09:0.04
-######
-significant_params: BoostType
 variables: HT_otherjets MT_htautau MT_l1 MT_l2 abs_dphi_htautauMET abs_dphi_l2MET dR_b1b2 dR_b1b2_boosted dR_l1l2 dR_l1l2Pt_htautau dphi_b1b2 dphi_hbbMET dphi_hbbhtautau dphi_l2MET p_zeta p_zetavisible pt_b1 pt_b2 pt_hbb pt_htautau pt_l1 pt_l1l2MET pt_l2
-use_mass_var: true
 
-
-
-[MediumMassRange]
-channels: eTau muTau
+[High : Common]
 mass_range: 450 900
-param_list: BoostType Grad
-param: V False
-param: UseBaggedBoost True
-param: nCuts 20
-param: NTrees 400
-param: shrinkage 0.1
-# NodePurityLimit
-# Pruning ??
-# param_range: nCuts 20:100:80
-######
-param_range: NTrees 100:900:400
-param_range: shrinkage 0.1:1.0:0.3
-param_range: BaggedSampleFraction 0.5:1.0:0.25
-param_range: MaxDepth 2:4:1
-param_range: MinNodeSize 0.01:0.09:0.04
-######
-significant_params: BoostType
 variables: HT_otherjets MT_htautau MT_l1 abs_dphi_htautauMET abs_dphi_l2MET costheta_METhbb dR_hbbhtautau dR_l1l2 dR_l1l2Pt_htautau dphi_b1b2 mass_top1 mass_top2 p_zeta p_zetavisible pt_b1 pt_b2 pt_hbb pt_htautau pt_l1 pt_l1l2MET pt_l2
-use_mass_var: true
-
-
 
 [FILES inputs]
 file: GluGluToRadionToHHTo2B2Tau_M-250_narrow.root    M250    1
